:recycle: Fix elements test
