+++ elements/lisk-chain/test/unit/chain.spec.ts
@@ -15,7 +15,7 @@
 
 import { NotFoundError, Batch, Database, InMemoryDatabase } from '@liskhq/lisk-db';
 import { codec } from '@liskhq/lisk-codec';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { Chain } from '../../src/chain';
 import { CurrentState, StateStore } from '../../src/state_store';
 import { createValidDefaultBlock, defaultNetworkIdentifier } from '../utils/block';
