+++ elements/lisk-chain/test/unit/transactions.spec.ts
@@ -11,7 +11,7 @@
  *
  * Removal or modification of this copyright notice is prohibited.
  */
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { Transaction } from '../../src/transaction';
 
 describe('blocks/transactions', () => {
