+++ elements/lisk-client/test/lisk-cryptography/ed.spec.ts
@@ -16,16 +16,21 @@
 import { cryptography } from '../../src';
 
 const {
-	signMessageWithPassphrase,
-	verifyMessageWithPublicKey,
-	printSignedMessage,
-	signAndPrintMessage,
-	signData,
-	signDataWithPassphrase,
-	signDataWithPrivateKey,
-	verifyData,
-	digestMessage,
-	createMessageTag,
+	ed: {
+		signMessageWithPassphrase,
+		verifyMessageWithPublicKey,
+		printSignedMessage,
+		signAndPrintMessage,
+		signData,
+		signDataWithPassphrase,
+		signDataWithPrivateKey,
+		verifyData,
+		getKeyPairFromPhraseAndPath,
+		getPublicKeyFromPrivateKey,
+	},
+	utils: {
+		createMessageTag,
+	},
 } = cryptography;
 
 const makeInvalid = (buffer: Buffer): Buffer => {
@@ -38,6 +43,7 @@ const makeInvalid = (buffer: Buffer): Buffer => {
 const changeLength = (buffer: Buffer): Buffer => Buffer.concat([Buffer.from('00', 'hex'), buffer]);
 
 describe('sign', () => {
+	const MAX_UINT32 = 4294967295;
 	const tag = createMessageTag('TST');
 	const networkIdentifier = Buffer.from('a5df2ed79994178c10ac168d6d977ef45cd525e95b7a8624', 'hex');
 	const defaultPassphrase = 'minute omit local rare sword knee banner pair rib museum shadow juice';
@@ -70,45 +76,6 @@ ${defaultSignature}
 			publicKey: Buffer.from(defaultPublicKey, 'hex'),
 			signature: Buffer.from(defaultSignature, 'hex'),
 		};
-	});
-
-	describe('#digestMessage', () => {
-		const strGenerator = (len: number, chr: string): string => chr.repeat(len);
-
-		it('should create message digest for message with length = 0', () => {
-			const msgBytes = digestMessage('');
-			const expectedMessageBytes = Buffer.from(
-				'3fdb82ac2a879b647f4f27f3fbd1c27e0d4e278f830b76295604035330163b79',
-				'hex',
-			);
-			expect(msgBytes).toEqual(expectedMessageBytes);
-		});
-		it('should create message digest for message in length range 1 - 253', () => {
-			const msgBytes = digestMessage(strGenerator(250, 'a'));
-			const expectedMessageBytes = Buffer.from(
-				'12832c687d950513aa5db6198b84809eb8fd7ff1c8963dca48ea57278523ec67',
-				'hex',
-			);
-			expect(msgBytes).toEqual(expectedMessageBytes);
-		});
-		it('should create message digest for message in length range 254 - 65536', () => {
-			const msgBytes = digestMessage(strGenerator(65535, 'a'));
-			const expectedMessageBytes = Buffer.from(
-				'73da94220312e71eb5c55c94fdddca3c06a6c18cb74a4a4a2cee1a82875c2450',
-				'hex',
-			);
-			expect(msgBytes).toEqual(expectedMessageBytes);
-		});
-		it('should create message digest for message in length range 65537 - 4294967296', () => {
-			const msgBytes = digestMessage(strGenerator(6710886, 'a'));
-			const expectedMessageBytes = Buffer.from(
-				'7c51817b5c31c4d04e9ffcf2e78859d6522b124f218c789a8f721b5f3e6b295d',
-				'hex',
-			);
-			expect(msgBytes).toEqual(expectedMessageBytes);
-		});
-		// highest range (length > 4294967296) is not practical to test
-		// but it is covered by `varuint-bitcoin` library
 	});
 
 	describe('#signMessageWithPassphrase', () => {
@@ -237,6 +204,69 @@ ${defaultSignature}
 				Buffer.from(defaultPublicKey, 'hex'),
 			);
 			expect(verification).toBe(true);
+		});
+	});
+
+	describe('getKeyPairFromPhraseAndPath', () => {
+		const passphrase =
+			'target cancel solution recipe vague faint bomb convince pink vendor fresh patrol';
+		it('should get keypair from valid phrase and path', async () => {
+			const privateKey = await getKeyPairFromPhraseAndPath(
+				passphrase,
+				`m/44'/134'/0'`,
+			);
+			const publicKey = getPublicKeyFromPrivateKey(privateKey);
+			expect(publicKey.toString('hex')).toBe(
+				'c6bae83af23540096ac58d5121b00f33be6f02f05df785766725acdd5d48be9d',
+			);
+			expect(privateKey.toString('hex')).toBe(
+				'c465dfb15018d3aef0d94d411df048e240e87a3ec9cd6d422cea903bfc101f61c6bae83af23540096ac58d5121b00f33be6f02f05df785766725acdd5d48be9d',
+			);
+		});
+
+		it('should fail for empty string path', async () => {
+			await expect(getKeyPairFromPhraseAndPath(passphrase, '')).rejects.toThrow(
+				'Invalid path format',
+			);
+		});
+
+		it('should fail if path does not start with "m"', async () => {
+			await expect(getKeyPairFromPhraseAndPath(passphrase, `/44'/134'/0'`)).rejects.toThrow(
+				'Invalid path format',
+			);
+		});
+
+		it('should fail if path does not include at least one "/"', async () => {
+			await expect(getKeyPairFromPhraseAndPath(passphrase, 'm441340')).rejects.toThrow(
+				'Invalid path format',
+			);
+		});
+
+		it('should fail for path with invalid segment', async () => {
+			await expect(
+				getKeyPairFromPhraseAndPath(
+					passphrase,
+					`m//134'/0'`, // should be number with or without ' between every back slash
+				),
+			).rejects.toThrow('Invalid path format');
+		});
+
+		it('should fail for path with invalid characters', async () => {
+			await expect(getKeyPairFromPhraseAndPath(passphrase, `m/a'/134b'/0'`)).rejects.toThrow(
+				'Invalid path format',
+			);
+		});
+
+		it('should fail for path with non-sanctioned special characters', async () => {
+			await expect(getKeyPairFromPhraseAndPath(passphrase, `m/4a'/#134b'/0'`)).rejects.toThrow(
+				'Invalid path format',
+			);
+		});
+
+		it(`should fail for path with segment greater than ${MAX_UINT32} / 2`, async () => {
+			await expect(
+				getKeyPairFromPhraseAndPath(passphrase, `m/44'/134'/${MAX_UINT32}'`),
+			).rejects.toThrow('Invalid path format');
 		});
 	});
 });
