+++ elements/lisk-client/test/lisk-cryptography/utils.spec.ts
@@ -15,7 +15,15 @@
 
 import { cryptography } from '../../src';
 
-const { bufferToHex, hexToBuffer, intToBuffer } = cryptography;
+const { utils: {
+	generateHashOnionSeed,
+	bufferToHex,
+	hexToBuffer,
+	intToBuffer,
+	getNetworkIdentifier,
+	hashOnion,
+	hash: hashFunction,
+}, } = cryptography;
 
 describe('buffer', () => {
 	const defaultBuffer = Buffer.from('\xe5\xe4\xf6');
@@ -88,7 +96,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt8(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 1 byte buffer when size=1, endian=little', () => {
@@ -99,7 +107,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt8(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 2 bytes big endian buffer when size=2, endian=big', () => {
@@ -110,7 +118,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt16BE(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 2 bytes little endian buffer when size=2, endian=little', () => {
@@ -121,7 +129,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt16LE(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 4 bytes big endian buffer when size=4, endian=big', () => {
@@ -132,7 +140,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt32BE(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 4 bytes little endian buffer when size=4, endian=little', () => {
@@ -143,7 +151,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt32LE(value, 0);
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 4 bytes big endian buffer when no size or endian is given', () => {
@@ -153,7 +161,7 @@ describe('buffer', () => {
 			const expectedBuffer = Buffer.alloc(size);
 			expectedBuffer.writeInt32BE(value, 0);
 
-			expect(utils.intToBuffer(value, size)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 8 bytes big endian buffer when size=8, endian=big', () => {
@@ -163,7 +171,7 @@ describe('buffer', () => {
 
 			const expectedBuffer = Buffer.from('00cebcaa8d34153d', 'hex');
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 8 bytes little endian buffer when size=8, endian=little', () => {
@@ -173,7 +181,7 @@ describe('buffer', () => {
 
 			const expectedBuffer = Buffer.from('3d15348daabcce00', 'hex');
 
-			expect(utils.intToBuffer(value, size, endian)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size, endian)).toEqual(expectedBuffer);
 		});
 
 		it('should convert a integer to a 8 bytes big endian buffer when size=8 and endian is not given', () => {
@@ -182,7 +190,110 @@ describe('buffer', () => {
 
 			const expectedBuffer = Buffer.from('00cebcaa8d34153d', 'hex');
 
-			expect(utils.intToBuffer(value, size)).toEqual(expectedBuffer);
+			expect(intToBuffer(value, size)).toEqual(expectedBuffer);
+		});
+	});
+
+	describe('hash onion', () => {
+		describe('#generateHashOnionSeed', () => {
+			it('should generate a random buffer', () => {
+				const seed1 = generateHashOnionSeed().toString('hex');
+				const seed2 = generateHashOnionSeed().toString('hex');
+
+				expect(seed1).not.toEqual(seed2);
+			});
+
+			it('should generate a random buffer with 16 bytes', () => {
+				const seed = generateHashOnionSeed();
+				expect(seed).toHaveLength(16);
+			});
+		});
+
+		describe('#hashOnion', () => {
+			let seed: Buffer;
+			let hashOnionBuffers: ReadonlyArray<Buffer>;
+
+			beforeEach(() => {
+				seed = generateHashOnionSeed();
+				hashOnionBuffers = hashOnion(seed);
+			});
+
+			it('should return 1001 hash onion hashes checkpoints by default', () => {
+				expect(hashOnionBuffers).toHaveLength(1001);
+			});
+
+			it('should return hash onion hashes which includes seed as the last element', () => {
+				expect(hashOnionBuffers[1000]).toEqual(seed);
+			});
+
+			it('should be able to calculate the checkpoint from another checkpoint', () => {
+				const firstDistanceHashes = hashOnion(hashOnionBuffers[1].slice(), 1000, 1);
+				expect(firstDistanceHashes[0]).toEqual(hashOnionBuffers[0]);
+				expect(firstDistanceHashes[1000]).toEqual(hashOnionBuffers[1]);
+			});
+		});
+	});
+
+	describe('hash', () => {
+		describe('#hash', () => {
+			const defaultText = 'text123*';
+			let arrayToHash: ReadonlyArray<number>;
+			let defaultHash: Buffer;
+
+			beforeEach(() => {
+				defaultHash = Buffer.from(
+					'7607d6792843d6003c12495b54e34517a508d2a8622526aff1884422c5478971',
+					'hex',
+				);
+				arrayToHash = [1, 2, 3];
+			});
+
+			it('should generate a sha256 hash from a Buffer', () => {
+				const testBuffer = Buffer.from(defaultText);
+				const hash = hashFunction(testBuffer);
+				expect(hash).toEqual(defaultHash);
+			});
+
+			it('should generate a sha256 hash from a utf8 string', () => {
+				const hash = hashFunction(defaultText, 'utf8');
+				expect(hash).toEqual(defaultHash);
+			});
+
+			it('should generate a sha256 hash from a hex string', () => {
+				const testHex = Buffer.from(defaultText).toString('hex');
+				const hash = hashFunction(testHex, 'hex');
+				expect(hash).toEqual(defaultHash);
+			});
+
+			it('should throw on unknown format when trying a string with format "utf32"', () => {
+				expect(hashFunction.bind(null, defaultText, 'utf32')).toThrow(
+					'Unsupported string format. Currently only `hex` and `utf8` are supported.',
+				);
+			});
+
+			it('should throw on unknown format when using an array', () => {
+				expect(hashFunction.bind(null, arrayToHash as any)).toThrow(
+					'Unsupported data:1,2,3 and format:undefined. Currently only Buffers or hex and utf8 strings are supported.',
+				);
+			});
+		});
+
+		describe('#getNetworkIdentifier', () => {
+			const genesisBlockID = Buffer.from(
+				'ed14889723f24ecc54871d058d98ce91ff2f973192075c0155ba2b7b70ad2511',
+				'hex',
+			);
+			const communityIdentifier = 'LISK';
+			const expectedHash = Buffer.from(
+				'6f201e72e20571b93ed42470caa94af1ace79dc9930ab5bb144ddd5df5753e73',
+				'hex',
+			);
+
+			it('should generate a sha256 hash from genesis block transaction root and community identifier', () => {
+				const networkIdentifier = getNetworkIdentifier(genesisBlockID, communityIdentifier);
+
+				expect(networkIdentifier).toEqual(expectedHash);
+			});
 		});
 	});
 });
