+++ framework/test/unit/modules/dpos_v2/commands/unlock.spec.ts
@@ -13,7 +13,7 @@
  */
 
 import { BlockHeader, Transaction } from '@liskhq/lisk-chain';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import * as testing from '../../../../../src/testing';
 import { UnlockCommand } from '../../../../../src/modules/dpos_v2/commands/unlock';
 import {
