+++ framework/test/unit/modules/dpos_v2/module.spec.ts
@@ -13,7 +13,7 @@
  */
 
 import { BlockAssets } from '@liskhq/lisk-chain';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { when } from 'jest-when';
 import { codec } from '@liskhq/lisk-codec';
 import { GenesisConfig } from '../../../../src/types';
