+++ framework/test/unit/modules/interoperability/mainchain/commands/state_recovery.spec.ts
@@ -14,7 +14,7 @@
 
 import { Transaction } from '@liskhq/lisk-chain';
 import { codec } from '@liskhq/lisk-codec';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { sparseMerkleTree } from '@liskhq/lisk-tree';
 import { CommandExecuteContext, CommandVerifyContext } from '../../../../../../src';
 import { BaseCCCommand } from '../../../../../../src/modules/interoperability/base_cc_command';
