+++ framework/test/unit/modules/interoperability/mainchain/store.spec.ts
@@ -12,7 +12,7 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { when } from 'jest-when';
 import { testing } from '../../../../../src';
 import {
