+++ framework/test/unit/modules/token/cc_commands/cc_forward.spec.ts
@@ -13,7 +13,7 @@
  */
 
 import { codec } from '@liskhq/lisk-codec';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { TokenAPI } from '../../../../../src/modules/token/api';
 import { CCForwardCommand } from '../../../../../src/modules/token/cc_commands/cc_forward';
 import {
