+++ framework/test/unit/state_machine/event_queue.spec.ts
@@ -13,7 +13,7 @@
  */
 
 import { EVENT_MAX_EVENT_SIZE_BYTES } from '@liskhq/lisk-chain';
-import { getRandomBytes, intToBuffer } from '@liskhq/lisk-cryptography';
+import { utils } from '@liskhq/lisk-cryptography';
 import { EventQueue } from '../../../src/state_machine/event_queue';
 
 describe('EventQueue', () => {
